rom2ram updated
ROM2RAM is a program that, when loaded via XModem, enables debugging info for the CF device driver routines. It shows LBA and buffers for every Read and Write actions from and to the CF-card. It also shows the result from that action (OK or error).
It does so by copying BIOS to RAM and then patch it.