FGC ROM: v0.21 added
These files build the latest FGC eprom version from Joerg Walke, 14-05-2025. A binary file-compare showed no differences for the first 16 KB of the .bin file, highest 16 KB are not build yet.
Note that this version does NOT run yet on the JC2 v4.1 system, it only served as a reference and baseline for future revisions.